+++ src/Generator/MethodGenerator.php
@@ -29,6 +29,16 @@ class MethodGenerator extends AbstractMemberGenerator
     protected $body = null;
 
     /**
+     * @var null|TypeGenerator
+     */
+    private $returnType;
+
+    /**
+     * @var bool
+     */
+    private $returnsReference = false;
+
+    /**
      * @param  MethodReflection $reflectionMethod
      * @return MethodGenerator
      */
@@ -39,6 +49,7 @@ class MethodGenerator extends AbstractMemberGenerator
 
         $method->setSourceContent($reflectionMethod->getContents(false));
         $method->setSourceDirty(false);
+        $method->setReturnType(self::extractReturnTypeFromMethodReflection($reflectionMethod));
 
         if ($reflectionMethod->getDocComment() != '') {
             $method->setDocBlock(DocBlockGenerator::fromReflection($reflectionMethod->getDocBlock()));
@@ -56,7 +67,7 @@ class MethodGenerator extends AbstractMemberGenerator
 
         $method->setInterface($declaringClass->isInterface());
         $method->setStatic($reflectionMethod->isStatic());
-
+        $method->setReturnsReference($reflectionMethod->returnsReference());
         $method->setName($reflectionMethod->getName());
 
         foreach ($reflectionMethod->getParameters() as $reflectionParameter) {
@@ -259,6 +270,32 @@ class MethodGenerator extends AbstractMemberGenerator
     }
 
     /**
+     * @param string|null
+     *
+     * @return MethodGenerator
+     */
+    public function setReturnType($returnType = null)
+    {
+        $this->returnType = null === $returnType
+            ? null
+            : TypeGenerator::fromTypeString($returnType);
+
+        return $this;
+    }
+
+    /**
+     * @param bool $returnsReference
+     *
+     * @return MethodGenerator
+     */
+    public function setReturnsReference($returnsReference)
+    {
+        $this->returnsReference = (bool) $returnsReference;
+
+        return $this;
+    }
+
+    /**
      * @return string
      */
     public function generate()
@@ -282,7 +319,9 @@ class MethodGenerator extends AbstractMemberGenerator
 
         $output .= $this->getVisibility()
             . (($this->isStatic()) ? ' static' : '')
-            . ' function ' . $this->getName() . '(';
+            . ' function '
+            . ($this->returnsReference ? '& ' : '')
+            . $this->getName() . '(';
 
         $parameters = $this->getParameters();
         if (!empty($parameters)) {
@@ -294,6 +333,10 @@ class MethodGenerator extends AbstractMemberGenerator
         }
 
         $output .= ')';
+
+        if ($this->returnType) {
+            $output .= ' : ' . $this->returnType->generate();
+        }
 
         if ($this->isAbstract()) {
             return $output . ';';
@@ -318,5 +361,29 @@ class MethodGenerator extends AbstractMemberGenerator
     public function __toString()
     {
         return $this->generate();
+    }
+
+    /**
+     * @param MethodReflection $methodReflection
+     *
+     * @return null|string
+     */
+    private static function extractReturnTypeFromMethodReflection(MethodReflection $methodReflection)
+    {
+        $returnType = method_exists($methodReflection, 'getReturnType')
+            ? $methodReflection->getReturnType()
+            : null;
+
+        if (! $returnType) {
+            return null;
+        }
+
+        $returnTypeString = (string) $returnType;
+
+        if ('self' === strtolower($returnType)) {
+            return $methodReflection->getDeclaringClass()->getName();
+        }
+
+        return $returnTypeString;
     }
 }