+++ test/Reflection/ParameterReflectionTest.php
@@ -10,6 +10,9 @@
 namespace ZendTest\Code\Reflection;
 
 use Zend\Code\Reflection;
+use ZendTest\Code\TestAsset\ClassTypeHintedClass;
+use ZendTest\Code\TestAsset\DocBlockOnlyHintsClass;
+use ZendTest\Code\TestAsset\InternalHintsClass;
 
 /**
  * @group      Zend_Reflection
@@ -41,13 +44,13 @@ class ParameterReflectionTest extends \PHPUnit_Framework_TestCase
     public function testTypeReturn($param, $type)
     {
         $parameter = new Reflection\ParameterReflection(['ZendTest\Code\Reflection\TestAsset\TestSampleClass5', 'doSomething'], $param);
-        $this->assertEquals($type, $parameter->getType());
+        $this->assertEquals($type, $parameter->detectType());
     }
 
     public function testCallableTypeHint()
     {
         $parameter = new Reflection\ParameterReflection(['ZendTest\Code\Reflection\TestAsset\CallableTypeHintClass', 'foo'], 'bar');
-        $this->assertEquals('callable', $parameter->getType());
+        $this->assertEquals('callable', $parameter->detectType());
     }
 
     public function paramTypeTestProvider()
@@ -58,6 +61,137 @@ class ParameterReflectionTest extends \PHPUnit_Framework_TestCase
             ['three','string'],
             ['array','array'],
             ['class','ZendTest\Code\Reflection\TestAsset\TestSampleClass']
+        ];
+    }
+
+    /**
+     * @group zendframework/zend-code#29
+     *
+     * @requires PHP 7.0
+     *
+     * @dataProvider reflectionHintsProvider
+     *
+     * @param string $className
+     * @param string $methodName
+     * @param string $parameterName
+     * @param string $expectedType
+     */
+    public function testGetType($className, $methodName, $parameterName, $expectedType)
+    {
+        $reflection = new Reflection\ParameterReflection(
+            [$className, $methodName],
+            $parameterName
+        );
+
+        $type = $reflection->getType();
+
+        self::assertInstanceOf(\ReflectionType::class, $type);
+        self::assertSame($expectedType, (string) $type);
+    }
+
+    /**
+     * @group zendframework/zend-code#29
+     *
+     * @requires PHP 7.0
+     *
+     * @dataProvider reflectionHintsProvider
+     *
+     * @param string $className
+     * @param string $methodName
+     * @param string $parameterName
+     * @param string $expectedType
+     */
+    public function testDetectType($className, $methodName, $parameterName, $expectedType)
+    {
+        $reflection = new Reflection\ParameterReflection(
+            [$className, $methodName],
+            $parameterName
+        );
+
+        // following is just due to an incompatibility between this test method and `testGetType`
+        if ('self' === $expectedType) {
+            $expectedType = $className;
+        }
+
+        self::assertSame($expectedType, $reflection->detectType());
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function reflectionHintsProvider()
+    {
+        return [
+            [InternalHintsClass::class, 'arrayParameter', 'foo', 'array'],
+            [InternalHintsClass::class, 'callableParameter', 'foo', 'callable'],
+            [InternalHintsClass::class, 'intParameter', 'foo', 'int'],
+            [InternalHintsClass::class, 'floatParameter', 'foo', 'float'],
+            [InternalHintsClass::class, 'stringParameter', 'foo', 'string'],
+            [InternalHintsClass::class, 'boolParameter', 'foo', 'bool'],
+            [ClassTypeHintedClass::class, 'selfParameter', 'foo', 'self'],
+            [ClassTypeHintedClass::class, 'classParameter', 'foo', ClassTypeHintedClass::class],
+            [ClassTypeHintedClass::class, 'otherClassParameter', 'foo', InternalHintsClass::class],
+            [ClassTypeHintedClass::class, 'closureParameter', 'foo', \Closure::class],
+            [ClassTypeHintedClass::class, 'importedClosureParameter', 'foo', \Closure::class],
+        ];
+    }
+
+    /**
+     * @group zendframework/zend-code#29
+     *
+     * @requires PHP 7.0
+     *
+     * @dataProvider docBlockHintsProvider
+     *
+     * @param string $className
+     * @param string $methodName
+     * @param string $parameterName
+     */
+    public function testGetTypeWithDocBlockOnlyTypes($className, $methodName, $parameterName)
+    {
+        $reflection = new Reflection\ParameterReflection(
+            [$className, $methodName],
+            $parameterName
+        );
+
+        self::assertNull($reflection->getType());
+    }
+
+    /**
+     * @group zendframework/zend-code#29
+     *
+     * @dataProvider docBlockHintsProvider
+     *
+     * @param string $className
+     * @param string $methodName
+     * @param string $parameterName
+     * @param string $expectedType
+     */
+    public function testDetectTypeWithDocBlockOnlyTypes($className, $methodName, $parameterName, $expectedType)
+    {
+        $reflection = new Reflection\ParameterReflection(
+            [$className, $methodName],
+            $parameterName
+        );
+
+        self::assertSame($expectedType, $reflection->detectType());
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function docBlockHintsProvider()
+    {
+        return [
+            [DocBlockOnlyHintsClass::class, 'arrayParameter', 'foo', 'array'],
+            [DocBlockOnlyHintsClass::class, 'callableParameter', 'foo', 'callable'],
+            [DocBlockOnlyHintsClass::class, 'intParameter', 'foo', 'int'],
+            [DocBlockOnlyHintsClass::class, 'floatParameter', 'foo', 'float'],
+            [DocBlockOnlyHintsClass::class, 'stringParameter', 'foo', 'string'],
+            [DocBlockOnlyHintsClass::class, 'boolParameter', 'foo', 'bool'],
+            [DocBlockOnlyHintsClass::class, 'selfParameter', 'foo', 'self'],
+            [DocBlockOnlyHintsClass::class, 'classParameter', 'foo', 'DocBlockOnlyHintsClass'],
+            [DocBlockOnlyHintsClass::class, 'otherClassParameter', 'foo', 'InternalHintsClass'],
         ];
     }
 }